+++ src/interceptors/add_is_logged_in_as_to_context.js
@@ -1,0 +1,13 @@
+/**
+   Add isLoggedInAs true to context.
+
+   Changes to `ctx`:
+
+   - add `isLoggedInAs`
+ */
+
+export default class AddIsLoggedInAsToContext {
+  enter(ctx) {
+    return { ...ctx, isLoggedInAs: true };
+  }
+}